% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveMCMC.R
\name{curveMCMC}
\alias{curveMCMC}
\title{Fit group-specific curves with MCMC}
\usage{
curveMCMC(df, t, y, Trt1, Trt2 = NA, GrpVar, numIntKnots = 10,
  sigmaBeta = 1e+05, Au = 1e+05, AU = 1e+05, Aeps = 1e+05,
  nwarm = 500, nsamp = 1000, odens = 1)
}
\arguments{
\item{df}{data frame.}

\item{t}{variable name in df which denotes time.}

\item{y}{response variable.}

\item{Trt1}{variable name of a factor in df which denotes the first treatment group. See in Details.}

\item{Trt2}{variable name of a factor in df which denotes the second treatment group. Set NA if no second treatment group.}

\item{GrpVar}{variable name in df which denotes the group variable.}

\item{numIntKnots}{number of interior knots.}

\item{sigmaBeta}{hyperparameter, variance in Normal disctribution of slope coefficient.}

\item{Au}{hyperparameter, scale in half Cauchy distribution of variance of random effects, assuming the random effects are iid.}

\item{AU}{hyperparameter, scale in half Cauchy distribution of variance of random intercept.}

\item{Aeps}{hyperparameter, scale in half Cauchy distribution of variance of residuals.}

\item{nwarm}{number of MCMC iterations as burn-in.}

\item{nsamp}{number of MCMC iterations.}

\item{number}{of MCMC iterations.}
}
\value{
An object contains MCMC samples is returned.

\item{MCMCsamples}{A list contains saved MCMC samples of slope, random effects, random intercepts, variance of random effects, variance of random intercept, variance of residuals.}

\item{data}{The data frame which is sorted by Trt1 and Trt2, and contains the new group coding by taking each combination of the treatment factors.}
}
\description{
Fit Bayesian penalized smoothing splines for longitudinal data for each level within a treatment group. Cubic O'Sullivan splines are used as basis functions in the linear mixed model formulation. Random intercepts are added for each group.
}
\details{
The current version only supports one or two treatments, both of which must be factors. If there is only one  treatment, then set \code{Trt2}=NA. An individual curve is fitted to each level of the (combination of) treatments.
The MCMC is fitted from stan, which interfaces with R using `rstan` package.
}
\examples{
library(fda)
library(reshape2)
#Berkeley growth data for 54 girls and 39 boys at 31 ages from 1-18 years old.
data(growth)
df_f<-melt(growth$hgtf)
df_m<-melt(growth$hgtm)
df<-rbind(df_f,df_m)
names(df)<-c("age","indiv","height")
df$gender<-c(rep("F",dim(df_f)[1]), rep("M",dim(df_m)[1]))
rs<-curveMCMC(df=df,t="age",y="height",Trt1="gender",Trt2=NA,GrpVar="indiv")

}
\references{
\insertRef{harezlak2018semiparametric}{Curvecompare}
}
\author{
Jiali Wang (\email{jiali.wang@data61.csiro.au})
}
