% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurve.R
\name{plotCurve}
\alias{plotCurve}
\title{Plot individual curves and their constrast curve}
\usage{
plotCurve(MCMCsamples = rs$MCMCsamples, compare1, compare2, numIntKnots,
  data = rs$data, plot.contrast, plot.indiv, credible_level = 0.95)
}
\arguments{
\item{MCMCsamples}{The object from \code{curveMCMC} which contains MCMC samples.}

\item{compare1}{Level in the treatment groups. If there is only 1 treatment group when fitting \code{curveMCMC}, it should be specified as compare1="a", where "a" is a factor level in Trt1. If there are two treatment groups when fitting  \code{curveMCMC}, it should be specified as compare1=c("a","b"), where "a" is a factor level in Trt1 and "b" is a factor level in Trt2.}

\item{compare2}{Level in the treatment groups. Similar to \code{compare1}.}

\item{numIntKnots}{number of interior knots. Must be the same as fitting \code{curveMCMC}.}

\item{data}{The data frame from fitting \code{curveMCMC}, which is sorted by Trt1 and Trt2.}

\item{plot.contrast}{plot two mean curves defined by \code{compare1} and \code{compare2}, and their contrast curve (curve1 minus curve2) with credible intervals. TRUE or FALSE.}

\item{plot.indiv}{plot two curves defined by \code{compare1} and \code{compare2} with respective credible intervals. TRUE or FALSE.}

\item{credible_level}{level of credible interval.}
}
\description{
Plot two curves corresponding to two levels from a single treatment group or from two treatment groups, and their contrast curve with credible intervals.
}
\examples{
#compare growth curves for females and males
plotCurve(MCMCsamples=rs$MCMCsamples,compare1="F",
compare2="M",numIntKnots=10,plot.contrast=TRUE,plot.indiv=TRUE,data=rs$data)

}
\author{
Jiali Wang (\email{jiali.wang@data61.csiro.au})
}
